requirements.txt
pipecat-ai[daily,elevenlabs,openai,silero,cartesia]
fastapi==0.115.6
uvicorn
python-dotenv
twilio
python-multipart
supabase
lancedb
sentence-transformers
